Move disk info creation to the recordings module
